requirements.txt
Flask>=2.0
PyYAML>=6.0
